requirements.txt
cvzone==1.6.1
numpy==2.0.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==10.4.0
protobuf==5.27.3
streamlit==1.37.1
